+++ example/views/dummy/index.js
@@ -6,6 +6,7 @@ import QuickCreate from './subviews/quick-create';
 import Link from   'components/link';
 import Pill from 'components/pill';
 import Banner from   'components/banner';
+import Toast from   'components/toast';
 
 import FinancesStore from './../../stores/finances';
 import UserActions from './../../actions/user';
@@ -17,6 +18,7 @@ import FinancesChart from './subviews/chart';
 import FinancesTable from './subviews/grid';
 import UserDialog from './subviews/user-dialog';
 import Flash from 'components/flash';
+import Message from 'components/message';
 
 class Finances extends React.Component {
 
@@ -38,26 +40,22 @@ class Finances extends React.Component {
     FinancesActions.financesFlashOpened();
   }
 
-  handleFlashEnd = (ev) => {
-    FinancesActions.financesFlashClosed();
-  }
-
   render() {
     let name = this.state.financesStore.get('name');
     let financesStore = this.state.financesStore;
 
     return (
       <div>
-        <Banner as="new" title="This is a title" message="This is my message." buttonAction={ function() {console.log('clicked')}}/>
+        <Banner as="warning" title="This is a title" message="This is my message." buttonAction={ function() {console.log('clicked')}}/>
 
         <div className="view-finances">
           <FinancesHistory />
 
           <Button onClick={ this.handleOnClick }>Edit My Details</Button>
 
-          <Pill type='warning'>Warning Pill</Pill>
-          <Pill ype='info'>Info Pill</Pill>
-          <Pill type='new'>New Pill</Pill>
+          <Pill as='warning'>Warning Pill</Pill>
+          <Pill as='info'>Info Pill</Pill>
+          <Pill as='new'>New Pill</Pill>
 
           <QuickCreate />
 
@@ -78,6 +76,8 @@ class Finances extends React.Component {
               data={ financesStore.get('chart_data') }
               balance={ financesStore.get('balance') } />
 
+            <Message>This is an example of a message.</Message>
+
             <FinancesTable
               data={ financesStore.get('line_items') }
               discount={ financesStore.get('discount') }
@@ -91,8 +91,13 @@ class Finances extends React.Component {
 
           <UserDialog />
 
-          <Flash onDismiss={ this.handleFlashEnd } message="Save Successful!" open={this.state.financesStore.get('displayFlash')} type="alert" />
+          <Flash onDismiss={ FinancesActions.dismissFlash } message="Save Successful!" open={this.state.financesStore.get('displayFlash')} as="success" />
         </div>
+
+        <Toast as="new" onDismiss={ FinancesActions.dismissToast } open={ this.state.financesStore.get('displayToast') }>
+          <strong>New Features</strong><br />
+          We have introduced new features, please see the <Link href="https://github.com/Sage/carbon/blob/master/CHANGELOG.md" target="_blank">changelog</Link> for more information.
+        </Toast>
       </div>
     );
   }