+++ src/components/confirm/confirm.js
@@ -16,8 +16,8 @@ import I18n from "i18n-js";
  *
  *   <Confirm
  *      title='Are you sure?"
- *      confirmHandler={ customConfirmHandler }
- *      cancelHandler={ customCancelHandler }
+ *      onConfirm={ customConfirmHandler }
+ *      onCancel={ customCancelHandler }
  *      open={ false }
  *    This is the content message
  *   </Confirm>
@@ -38,10 +38,10 @@ class Confirm extends Dialog {
     /**
      * A custom event handler when a confirmation takes place
      *
-     * @property confirmHandler
+     * @property onConfirm
      * @type {Function}
      */
-    confirmHandler: React.PropTypes.func.isRequired
+    onConfirm: React.PropTypes.func.isRequired
   }
 
   constructor() {
@@ -93,11 +93,11 @@ class Confirm extends Dialog {
     return (
       <div className='ui-confirm__buttons' >
         <div className='ui-confirm__button ui-confirm__no'>
-          <Button as='secondary' onClick={ this.props.cancelHandler }>{ cancelText() }</Button>
+          <Button as='secondary' onClick={ this.props.onCancel }>{ cancelText() }</Button>
         </div>
 
         <div className='ui-confirm__button ui-confirm__yes'>
-          <Button as='primary' onClick={ this.props.confirmHandler }>{ confirmText() }</Button>
+          <Button as='primary' onClick={ this.props.onConfirm }>{ confirmText() }</Button>
         </div>
       </div>
     );