+++ src/components/dialog/dialog.js
@@ -14,7 +14,7 @@ import Bowser from 'bowser';
  *
  * To render a Dialog:
  *
- *   <Dialog cancelHandler={ customEventHandler } />
+ *   <Dialog onCancel={ customEventHandler } />
  *
  * The component rendering the Dialog must pass down a prop of 'open' in order to open the dialog.
  *
@@ -31,10 +31,10 @@ class Dialog extends React.Component {
     /**
      * A custom close event handler
      *
-     * @property cancelHandler
+     * @property onCancel
      * @type {Function}
      */
-    cancelHandler: React.PropTypes.func.isRequired,
+    onCancel: React.PropTypes.func.isRequired,
 
     /**
      * Sets the open state of the dialog
@@ -70,9 +70,21 @@ class Dialog extends React.Component {
   getChildContext() {
     return {
       dialog: {
-        cancelHandler: this.props.cancelHandler
+        onCancel: this.props.onCancel
       }
     };
+  }
+
+  /**
+   * A lifecycle method to update the component on initialize
+   *
+   * @method componentDidMount
+   * @return {void}
+   */
+  componentDidMount() {
+    if (this.props.open) {
+      this.centerDialog();
+    }
   }
 
   /**
@@ -81,7 +93,7 @@ class Dialog extends React.Component {
    * @method componentDidUpdate
    * @return {void}
    */
-  componentDidUpdate = () => {
+  componentDidUpdate() {
     if (this.props.open && !this.listening) {
       this.centerDialog();
       this.listening = true;
@@ -103,7 +115,7 @@ class Dialog extends React.Component {
    */
   closeDialog = (ev) => {
     if (ev.keyCode === 27) {
-      this.props.cancelHandler();
+      this.props.onCancel();
     }
   }
 
@@ -211,7 +223,7 @@ class Dialog extends React.Component {
     return (
       <div ref="dialog" className={ dialogClasses }>
         { this.dialogTitle }
-        <Icon className="ui-dialog__close" type="close" onClick={ this.props.cancelHandler } />
+        <Icon className="ui-dialog__close" type="close" onClick={ this.props.onCancel } />
 
         <div className='ui-dialog__content'>
           { this.props.children }