+++ src/components/icon/__spec__.js
@@ -3,17 +3,39 @@ import TestUtils from 'react/lib/ReactTestUtils';
 import Icon from './icon';
 
 describe('Icon', () => {
-  let instance, span;
+  let instance, span, svg;
 
   describe('renderIcon', () => {
-    beforeEach(() => {
-      let warningInstance = TestUtils.renderIntoDocument(<Icon type='warning' />);
-      spyOn(warningInstance, 'renderWarningIcon');
+    it('calls the render warning icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='warning' />);
+      instance.renderIcon;
+      span = TestUtils.findRenderedDOMComponentWithTag(instance, 'span');
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--warning");
     });
 
-    it('calls the render warning icon method', () => {
-      warningInstance.renderIcon;
-      expect(warningInstance.renderWarningIcon).toHaveBeenCalled();
+    it('calls the render new icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='new' />);
+      instance.renderIcon;
+      span = TestUtils.findRenderedDOMComponentWithTag(instance, 'span');
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--new");
+    });
+
+    it('calls the render maintenance icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='maintenance' />);
+      instance.renderIcon;
+      span = TestUtils.findRenderedDOMComponentWithTag(instance, 'span');
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--maintenance");
+    });
+  });
+
+  describe('success', () => {
+    it('renders with an icon of tick', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='success' />);
+      span = TestUtils.findRenderedDOMComponentWithTag(instance, 'span');
+      expect(span.className).toEqual('icon-tick');
     });
   });