+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.3.0
 
-TBC
+## Handler Pattern
+
+Carbon now has a simple handler pattern implementation. For more information, see [the guide](https://github.com/Sage/carbon/blob/master/docs/guides/handlers.md).
 
 # 0.2.0